+++ src/hooks/__internal__/useStableCallback/useStableCallback.ts
@@ -1,0 +1,15 @@
+import { useRef, useCallback } from "react";
+
+// This is a hook that returns a memoized callback that is guaranteed to be the same reference between renders.
+// This is useful when a stable reference is required to prevent unnecessary re-renders.
+
+export function useStableCallback(callback?: (...args: unknown[]) => unknown) {
+  const ref = useRef(callback);
+  ref.current = callback;
+
+  const stableCallback = useCallback((...args) => ref.current?.(...args), []);
+
+  return callback ? stableCallback : undefined;
+}
+
+export default useStableCallback;